+++ src/main/java/com/team/berp/mrp/repository/EntityItemRepository.java
@@ -8,7 +8,7 @@ import com.team.berp.domain.Item;
 import com.team.berp.domain.Mrp;
 
 // JpaRepository<엔티티, PK 타입>을 상속받아, DB 접근 기본 기능 자동 제공
-public interface EntityItemRepository extends JpaRepository<Item, Integer> {
+public interface EntityItemRepository extends JpaRepository<Item, Long> {
     // JpaRepository를 상속하면 별도 코드 없이 기본 CRUD 메서드 자동 제공
     // 예: findAll(), findById(), save(), deleteById() 등
